EXP-2296-3 Table export: Merge export engine base classes to reduce complexity
