+++ platform-emf/src/main/java/com/softicar/platform/emf/data/table/export/implementation/html/TableExportHtmlEngine.java
@@ -9,7 +9,7 @@ import com.softicar.platform.emf.data.table.export.conversion.factory.TableExpor
 import com.softicar.platform.emf.data.table.export.conversion.factory.TableExportStrictNodeConverterFactory;
 import com.softicar.platform.emf.data.table.export.conversion.factory.TableExportTextOnlyNodeConverterFactory;
 import com.softicar.platform.emf.data.table.export.conversion.factory.configuration.TableExportNodeConverterFactoryConfiguration;
-import com.softicar.platform.emf.data.table.export.engine.AbstractTableExportColumnFilteringEngine;
+import com.softicar.platform.emf.data.table.export.engine.AbstractTableExportEngine;
 import com.softicar.platform.emf.data.table.export.engine.ITableExportEngine;
 import com.softicar.platform.emf.data.table.export.engine.configuration.TableExportEngineConfiguration;
 import com.softicar.platform.emf.data.table.export.engine.configuration.TableExportTableConfiguration;
@@ -31,13 +31,11 @@ import java.util.SortedMap;
 import java.util.TreeMap;
 import org.apache.commons.text.StringEscapeUtils;
 
-public class TableExportHtmlEngine extends AbstractTableExportColumnFilteringEngine<TableExportTypedNodeValue, HtmlRow, HtmlCell> {
+public class TableExportHtmlEngine extends AbstractTableExportEngine<TableExportTypedNodeValue, HtmlRow, HtmlCell> {
 
 	private static final Charset OUTPUT_FILE_CHARSET = StandardCharsets.UTF_8;
 	private static final Charset CSS_FILE_CHARSET = StandardCharsets.UTF_8;
 	private static final int NUM_TABLE_SPACER_ROWS = 2;
-
-//	private static final char OUTPUT_SEPARATOR = ' ';
 
 	private PrintWriter printWriter = null;
